+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migration of influxdb-client to TS
 - Migration of traditonalcheck.js to TS
 - Changed export of redisClient.ts
+- Migration of gptchatcompletion to TS
 
 ### Added
 - Puppeteer and test cases